Checks: Move classloader to models
It belongs there as it depends on appconf from the models being
initialized. Before this was triggered by import from the trans module,
but it is now gone.
