+++ weblate/checks/__init__.py
@@ -16,22 +16,5 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-from django.utils.functional import cached_property
-
-from weblate.utils.classloader import ClassLoader
 
 default_app_config = "weblate.checks.apps.ChecksConfig"
-
-
-class ChecksLoader(ClassLoader):
-    @cached_property
-    def source(self):
-        return {k: v for k, v in self.items() if v.source}
-
-    @cached_property
-    def target(self):
-        return {k: v for k, v in self.items() if v.target}
-
-
-# Initialize checks list
-CHECKS = ChecksLoader("CHECK_LIST")
