+++ weblate/trans/filter.py
@@ -23,7 +23,7 @@ from django.utils.text import format_lazy
 from django.utils.translation import gettext
 from django.utils.translation import gettext_lazy as _
 
-from weblate.checks import CHECKS
+from weblate.checks.models import CHECKS
 
 
 class FilterRegistry:
